[cdk] upgrade to 0.11
The whole payment subsystem has changed, minting/melting are now in the
process of being refactored to support multiple payment processor.
That means that structures are a bit different.

This commit removes the minting/melting parts from the wallet-aggregator
and sets nginx to redirect such requests to cdk-mintd directly.

Minor changes to adapt to changes in cdk structures