[bcr-wdc-key-service] refactor to keyset-per-day architecture
jumbo-refactor of the key-service that migrates its logic
from a keyset-per-ebill architecture to keyset-per-redemption date.

Previously, the key-service needed to have two separate keysDB, one for
temporary keysets generated during the quoting and minting process and
one for circulating `crsat`s,
We now no longer need such setup as there not longer exists such temporary
state where the keyset is generated for an ebill that is not yet
correctly endorsed to the wildcat.
Each keysetID is linked 1:1 to a redemption date and is enabled and
usable by default.
The minting operations are stored in a separate table and are filled by
a API call by the quote-service.